+++ src/generate-secret-key.c
@@ -1,10 +1,10 @@
 #include "keys.h"
 #include "modular-inverse.h"
 
-SecretKey generateSecretKey(int p, int q, int e) {
-    int number = (p - 1) * (q - 1);
-    int n = p * q;
-    int inverse = modularInverse(e, number);
+SecretKey generateSecretKey(long long p, long long q, long long e) {
+    long long number = (p - 1) * (q - 1);
+    long long n = p * q;
+    long long inverse = modularInverse(e, number);
 
     return (SecretKey){ n, inverse };
 }
