+++ src/modular-exponentiation.c
@@ -1,7 +1,7 @@
 #include "modular-exponentiation.h"
 
-int modularExponentiation(int base, int expo, int mod) {
-    int rest = 1;
+long long modularExponentiation(long long base, long long expo, long long mod) {
+    long long rest = 1;
 
     while (expo > 0) {
         if (expo % 2 == 1) {
