+++ src/modular-inverse.c
@@ -1,15 +1,15 @@
 #include "modular-inverse.h"
 #include "extended-euclides.h" 
 
-int modularInverse(int a, int m) {
-    int s, t;
-    int mdc = extendedEuclides(a, m, &s, &t);
+long long modularInverse(long long a, long long m) {
+    long long s, t;
+    long long mdc = extendedEuclides(a, m, &s, &t);
 
     if (mdc != 1) {
         return -1;
     }
 
-    int inverse = (s % m + m) % m;
+    long long inverse = (s % m + m) % m;
 
     if (inverse == 0) {
         return m;
