+++ types/decimal-to-binary.h
@@ -1,6 +1,6 @@
 #ifndef DECIMAL_TO_BINARY_H
 #define DECIMAL_TO_BINARY_H
 
-int decimalToBinary(int n, int binary[]);
+long long decimalToBinary(long long n, long long binary[]);
 
 #endif
